refactor(fmt): flatten formatter sources
